refactor: tokenManger, fetchWithAuth, authApi 파일 분리
